Rather bodgy fix for complete date
This allows PCRT to show the WO in the recently completed section by setting the readydate field. This method checks the is_on_site setting in config for specific states. The state ID and state information has to be forwarded to set_work_order_state to facilitate this.
This needs better tuning, and should support readydate.