feat: implement daily report + transactions log
Implements the daily report feature, this allows the system to show a report of how how it has been used on a specific day. Daily report can be triggered by a cron setting in the config or by the button on the dashboard.

Daily report also requires the newly added transactions log, which stores all transactions with the PCRT-Scan system (such as a scan, or action apply) in a local sqlite database. If transaction logging is enabled a new database will be created.

Transaction logging must be enabled to use daily report, if daily report is enabled but transactions are disabled, the server will fail to start.