feat-nyi: QR Code Action Buttons, testing
Adds a QR code to action buttons on work order view modal. This is to test applying actions using the scanner rather than interacting directly with PCRT-Scan.

Not yet implemented, the QR code is stored on the server as nyi.png but will be replaced with a dynamically generated QR code if this feature is implemented.

Temporary implementation serves to test the scanners ability to read from the LCD panel.